Handle exceptions in Service layer

+++ src/main/java/com/solvd/training/service/impl/DepartmentService.java
@@ -1,6 +1,8 @@
 package com.solvd.training.service.impl;
 
 import com.solvd.training.dao.impl.DepartmentDAO;
+import com.solvd.training.exceptions.NotFoundException;
+import com.solvd.training.model.Client;
 import com.solvd.training.model.Department;
 import com.solvd.training.service.IService;
 
@@ -14,17 +16,32 @@ public class DepartmentService implements IService<Department> {
     }
 
     @Override
-    public void update(int id, Department department) {
-        departmentDAO.update(id, department);
+    public void update(int id, Department department) throws NotFoundException {
+        Department foundDepartment = departmentDAO.find(id);
+        if (foundDepartment != null) {
+            departmentDAO.update(id, foundDepartment);
+        } else {
+            throw new NotFoundException("Department was not found");
+        }
     }
 
     @Override
-    public void delete(int id) {
-        departmentDAO.delete(id);
+    public void delete(int id) throws NotFoundException {
+        Department foundDepartment = departmentDAO.find(id);
+        if (foundDepartment != null) {
+            departmentDAO.delete(id);
+        } else {
+            throw new NotFoundException("Department was not found");
+        }
     }
 
     @Override
-    public Department find(int id){
-        return departmentDAO.find(id);
+    public Department find(int id) throws NotFoundException {
+        Department department = departmentDAO.find(id);
+        if (department != null) {
+            return department;
+        } else {
+            throw new NotFoundException("Department was not found");
+        }
     }
 }
