+++ src/main/java/com/solvd/training/service/impl/EmployeeService.java
@@ -1,6 +1,8 @@
 package com.solvd.training.service.impl;
 
 import com.solvd.training.dao.impl.EmployeeDAO;
+import com.solvd.training.exceptions.NotFoundException;
+import com.solvd.training.model.Department;
 import com.solvd.training.model.Employee;
 import com.solvd.training.service.IService;
 
@@ -14,17 +16,32 @@ public class EmployeeService implements IService<Employee>  {
     }
 
     @Override
-    public void update(int id, Employee employee) {
-        employeeDAO.update(id, employee);
+    public void update(int id, Employee employee) throws NotFoundException {
+        Employee foundEmployee = employeeDAO.find(id);
+        if (foundEmployee != null) {
+            employeeDAO.update(id, employee);
+        } else {
+            throw new NotFoundException("Employee was not found");
+        }
     }
 
     @Override
-    public void delete(int id) {
-        employeeDAO.delete(id);
+    public void delete(int id) throws NotFoundException {
+        Employee foundEmployee = employeeDAO.find(id);
+        if (foundEmployee != null) {
+            employeeDAO.delete(id);
+        } else {
+            throw new NotFoundException("Employee was not found");
+        }
     }
 
     @Override
-    public Employee find(int id){
-        return employeeDAO.find(id);
+    public Employee find(int id) throws NotFoundException {
+        Employee employee = employeeDAO.find(id);
+        if (employee != null) {
+            return employee;
+        } else {
+            throw new NotFoundException("Employee was not found");
+        }
     }
 }
