+++ src/main/java/com/solvd/training/service/impl/ProjectService.java
@@ -1,6 +1,9 @@
 package com.solvd.training.service.impl;
 
 import com.solvd.training.dao.impl.ProjectDAO;
+import com.solvd.training.exceptions.NotFoundException;
+import com.solvd.training.model.Department;
+import com.solvd.training.model.Employee;
 import com.solvd.training.model.Project;
 import com.solvd.training.service.IService;
 
@@ -14,17 +17,32 @@ public class ProjectService implements IService<Project> {
     }
 
     @Override
-    public void update(int id, Project project) {
-        projectDAO.update(id, project);
+    public void update(int id, Project project) throws NotFoundException {
+        Project foundProject = projectDAO.find(id);
+        if (foundProject != null) {
+            projectDAO.update(id, project);
+        } else {
+            throw new NotFoundException("Project was not found");
+        }
     }
 
     @Override
-    public void delete(int id) {
-        projectDAO.delete(id);
+    public void delete(int id) throws NotFoundException {
+        Project foundProject = projectDAO.find(id);
+        if (foundProject != null) {
+            projectDAO.delete(id);
+        } else {
+            throw new NotFoundException("Project was not found");
+        }
     }
 
     @Override
-    public Project find(int id){
-        return projectDAO.find(id);
+    public Project find(int id) throws NotFoundException {
+        Project project = projectDAO.find(id);
+        if (project != null) {
+            return project;
+        } else {
+            throw new NotFoundException("Project was not found");
+        }
     }
 }
