+++ src/main/java/com/solvd/training/service/impl/TaskService.java
@@ -1,6 +1,8 @@
 package com.solvd.training.service.impl;
 
 import com.solvd.training.dao.impl.TaskDAO;
+import com.solvd.training.exceptions.NotFoundException;
+import com.solvd.training.model.Project;
 import com.solvd.training.model.Task;
 import com.solvd.training.service.IService;
 
@@ -14,17 +16,32 @@ public class TaskService implements IService<Task> {
     }
 
     @Override
-    public void update(int id, Task task) {
-        taskDAO.update(id, task);
+    public void update(int id, Task task) throws NotFoundException {
+        Task foundTask = taskDAO.find(id);
+        if (foundTask != null) {
+            taskDAO.update(id, task);
+        } else {
+            throw new NotFoundException("Task was not found");
+        }
     }
 
     @Override
-    public void delete(int id) {
-        taskDAO.delete(id);
+    public void delete(int id) throws NotFoundException {
+        Task foundTask = taskDAO.find(id);
+        if (foundTask != null) {
+            taskDAO.delete(id);
+        } else {
+            throw new NotFoundException("Task was not found");
+        }
     }
 
     @Override
-    public Task find(int id){
-        return taskDAO.find(id);
+    public Task find(int id) throws NotFoundException {
+        Task task = taskDAO.find(id);
+        if (task != null) {
+            return task;
+        } else {
+            throw new NotFoundException("Task was not found");
+        }
     }
 }
